moved a file test
